Move collation-aware comparators to the types module
